Location Setting on change behavior restore
- To automatically update the UI when the location changes, we have to
update the weather AND restart the LoadManager
- Since the ForecastFragment will actually update the UI(because it
itself has the UI), it will have the procedure to do the . But to use
it, we have to go to the MainActivity
a. Right in MainActivities onCreate, we set the mLocation to be the
current setting (and use a tag for ForecastFragment)
b. We use its onResume method to do the calling if there is any  change
in setting

For now, we’re removing ForecastFragment’s onStart to avoid excessive
updates, and instead relying on the refresh button. We’ll learn to
schedule updates later.
strings.xml had a mistake, the values for metric and imperial are
supposed to correspond to those declared in arrays.xml

A BUG (in the list UI): It seems that since the weather temperature
units update mechanism (on settings change) is, unlike the location
update(which is changed from the ground up by requesting new data
itself via FetchWeatherTask), a purely UI phenomenon (for efficiency)
carried out by the CursorAdapter, the updating is tied to the list
items’ life. Hence, they don’t get updated unless scrolled and forced
to restart.