Fixed ImageView nullPointer bug
A critical mistake in a place unlooked for: had forgotten to define
icon’s id in the special today layout.
This left the very first view inflated by newView without a found image
in the holder resulting in a null pointer for iconView at the very
beginning.
The remaining normal views, while having the proper image id, have no
chance to correct the null reference now the we “hold” the found
views(in this case, even the unfound one)

Also the significance of @+id over @id:
“You need to use @+id when you are defining your own Id for a View.”