Add the downloaded icon folders to the res directory
Weather icons for all screen sizes are in their respectable drawable
folders. The new launcher icons are in the mipmap folders, where unlike
the drawables, the unwanted sizes won’t be stripped away so that sizes
different to the current screen sizes can be used if needed(Eg. the
launcher screen)