Make the tablet UI: Preparation
Loads of stuff:

for 2pane:
- delete the not needed values820 folder and add a layout-sw600 needed
to hold the tablet layout file: another activity_main.xml
- This will kick off if the device has smaller width >=600dp. It’s just
a linear layout(new divider prop) with 2 things:
   - an actual [fragment] which is none other our forecast_fragment
   - a container(FrameLayout) for the fresh detail fragments which’ll
be added to it dynamically. Better to have its id same as the container
in the phone layout(a FrameL held in an activity: activity_detail):
‘weather_detail_container’. Don’t  forget to update these changed ids
where they are used(in onCreate of Activities)
“In my case, I’ve removed that part, to have no more fragment
reinflation hassles”
- Handle the case of no URI in intent in the detailFragment; no loader
creation. Possible for when DFragment is in directly in the Main
activity, which has no URI in its own intent.

1pane:
- Make the forecast fragment static here as well: a lone fragment
instead of a Frame in activity_main
- Thus we no more create this fragment, we don’t need it’s tag in
MainActivity’s onResume. Find it plainly by its id. Make a tag for what
we ‘will’ have to create though: the Detail Fragment

Different inits for 1 or 2 pane:
- Following from above, remove the dynamic adding(the saved instance
null part) in MainActivity. Instead, for this condition, do a
transaction to REPLACE(not add) a ‘Detail’ Fragment, ONLY if it’s a
2pane(found by looking for the existence of the
weather_detail_container specific to 2pane)

First Tablet Layout blues:
Tablet AVD loads up phone UI. The skin definition AVD configuration
seems to be the cause: http://stackoverflow.com/a/32722831 As a result
noticed a 'skin not large enough to view entire screen' warning when
the skin definition is a default 'nexus_7_2013’.
That may explain the emulator screen being apparently less than 600dp
and hence showing the phone layout. However, it looks like the warning
is present in all default configured AVDs, nevertheless setting the
skin to 'No Skin' resolves it.