FIX the share inflation
Fixing the fragment inflating twice had the side effect of bungling the
share button, because it was inflated in the DetailActivity(so that it
is done only once) but this Activity will not be present in 2pane.
To account for both 1pane and 2pane, do all of the menu layout work in
the onCreateOptionsMenu of the fragment itself.
The cause was always that this method is called twice, hence we have to
overtime clear menu before inflating.