Use a service to refresh instead of the old Asynctask
Service: A new component - a new package, a new class, and a new
mention in the manifest.
Give the role of FetchWeatherTask to this class, overriding
onHandleIntent instead of doInBackground.
Notice the location passing: the location member variable string is
set[putExtra] and used[getStringExtra] while sending and receiving the
explicit intent respectively.