Call service with an alarm
A new broadcast receiver in the manifest and as an inner class, which
will now do the job of calling the service.
updateWeather on the other hand, will send it a pending intent wrapping
an explicit intent having the location, but only when the alarm fires
off, i.e. when the alarm conditions set on the alarm manager are met.